some more improvemnets
